+++ profiling/src/profile/internal/mod.rs
@@ -1,8 +1,10 @@
+mod function;
 mod label;
 mod line;
 mod location;
 mod value_type;
 
+pub use function::{Function, FunctionId};
 pub use label::{Label, LabelValue};
 pub use line::Line;
 pub use location::{Location, LocationId};
